+++ src/ast/passes/printer.h
@@ -45,6 +45,7 @@ public:
   void visit(If &if_node);
   void visit(Unroll &unroll);
   void visit(While &while_block);
+  void visit(Range &range);
   void visit(For &for_loop);
   void visit(Config &config);
   void visit(Jump &jump);