+++ CHANGELOG.md
@@ -51,6 +51,8 @@ and this project adheres to
   - [#4129](https://github.com/bpftrace/bpftrace/pull/4129)
 - Warn on discarded return values for specific builtin functions
   - [#4191](https://github.com/bpftrace/bpftrace/pull/4191)
+- For loops now support `..` ranges
+  - [#4126](https://github.com/bpftrace/bpftrace/pull/4126)
 #### Changed
 - `-p` CLI flag now applies to all probes (except BEGIN/END)
   - [#3800](https://github.com/bpftrace/bpftrace/pull/3800)